+++ odoo/local-src/specific_stock/models/__init__.py
@@ -9,3 +9,4 @@ from . import stock_inventory
 from . import stock_pack_operation
 from . import stock_quant_package
 from . import procurement_order
+from . import stock_scrap