+++ odoo/local-src/connector_esb/models/sale_order/common.py
@@ -2,7 +2,6 @@
 # Copyright 2017 Camptocamp SA
 # License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl.html)
 
-
 import logging
 
 from psycopg2 import IntegrityError