+++ odoo/local-src/specific_zetes/tests/test_assignment_reserve.py
@@ -29,6 +29,30 @@ class TestAssignemnt(ZetesReserveTest):
         self.assertEqual(result.respCode, str(constants.RESPONSE_CODE_OK))
         self.assertEqual(result.Usf09, '1')  # Nbr of lines
 
+        # Try to create an inventory for the product 1
+        # This picking should be not available
+        inventory = self.env['stock.inventory'].create({
+            'name': 'Test',
+            'filter': 'partial',
+        })
+        inventory.line_ids.create({
+            'inventory_id': inventory.id,
+            'product_id': self.product_1.id,
+            'product_qty': 20,
+            'location_id': self.env.ref('stock.stock_location_stock').id
+        })
+        # Start the inventory
+        inventory.action_start()
+        result_str = domain.requ(request_params)
+        result = self.format_result(result_str)
+        self.assertEqual(result.respCode, str(constants.RESPONSE_CODE_ERROR))
+
+        # Now validate the inventory
+        inventory.action_done()
+        result_str = domain.requ(request_params)
+        result = self.format_result(result_str)
+        self.assertEqual(result.respCode, str(constants.RESPONSE_CODE_OK))
+
     def test_02_requ_assignment(self):
         report_query = """
         SELECT report.id
@@ -38,13 +62,41 @@ class TestAssignemnt(ZetesReserveTest):
             ON stock_location.picking_zone_id = picking_zone.id
         WHERE report.product_id = %s
           AND picking_zone.code = %s
+          AND NOT EXISTS (SELECT 1
+                          FROM stock_inventory_line AS sil
+                            INNER JOIN stock_inventory AS si
+                              ON sil.inventory_id = si.id
+                          WHERE si.state = 'confirm'
+                          AND sil.location_id = report.location_id)
         ORDER BY report.refill_priority
         LIMIT 1
         """
+
+        # Try to create an inventory for the product 1
+        # No product should be available
+        inventory = self.env['stock.inventory'].create({
+            'name': 'Test',
+            'filter': 'partial',
+        })
+        inventory.line_ids.create({
+            'inventory_id': inventory.id,
+            'product_id': self.product_1.id,
+            'product_qty': 20,
+            'location_id': self.reserve_medoc.id
+        })
+        # Start the inventory
+        inventory.action_start()
         self.env.cr.execute(report_query, (self.product_1.id,
                                            self.picking_zone_medoc.code))
         result = self.env.cr.fetchone()
+        self.assertFalse(result)
 
+        # Now validate the inventory
+        # The product 1 should now be available
+        inventory.action_done()
+        self.env.cr.execute(report_query, (self.product_1.id,
+                                           self.picking_zone_medoc.code))
+        result = self.env.cr.fetchone()
         self.assertTrue(result)
         report_id = result[0]