+++ HISTORY.rst
@@ -8,7 +8,7 @@ Unreleased
 
 **Features and Improvements**
 
-* ALCN-1116: Add the module CSV File Export (to manage CSV Exports)
+* ALCN-1116: Add the module CSV File Export and CSV File Import
 
 **Bugfixes**