+++ HISTORY.rst
@@ -14,6 +14,7 @@ Unreleased
 * Limit multiple job creation for the same update on a sale order.
 * Install monitoring_status. Previously it has been manually installed and we want to keep it
 * ALCN-1379: Remove taxes from SER-705008 accounting products
+* ALCN-1403: Zetes: ignore pickings who contain inventoried products
 
 **Bugfixes**